requirements.txt
# PSTIF-WRO

torch>=1.12.0
torch-geometric>=2.1.0

numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

pandas>=1.3.0

matplotlib>=3.4.0
seaborn>=0.11.0

pyyaml>=5.4.0
tqdm>=4.62.0

# geomloss>=0.2.4
# POT>=0.8.0
